添加商品进购物车

+++ gmall-cart/src/main/java/com/study/gmall/pojo/Cart.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Data
 public class Cart {
 
-    private Integer skuId;
+    private Long skuId;
     private String title;
     private String defaultImage;
     private BigDecimal price;
